pycon_project/requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.

django-email-confirmation==0.2.dev2
django-timezones==0.2.dev1
django-ajax-validation==0.1.3
django-oauth-access==0.1.dev6
pytz==2010h
django-openid==0.2.3.dev1
python-openid==2.2.5
oauth2==1.1.3
biblion==0.1.dev6
creole==1.1
django-atomformat==0.1.1.dev1
Pygments==1.3.1
django-mailer==0.2a1.dev3
docutils==0.6
django-wakawaka==0.4.dev4
django-markitup==0.6.1
django-fixture-generator==0.2.0
nashvegas==0.2a1.dev4
idios==0.2a1.dev1